+++ src/pages/addnews/[...camp].tsx
@@ -1,4 +1,4 @@
-import Add from "../../components/ComponentPages/News/Add";
+import Add from "../../components/ComponentPages/News/AddOrEdit";
 import SideBarNoFilter from "../../components/ComponentPages/Home/SideBarNoFilter";
 import Layout from "../../hoc/layout";
